requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0
pandas==2.1.0 